backend/requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.5
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
